Started separation in modules backend-bikesurbanfleets-common --> Common classes and utils backend-bikesurbanfleets-core --> Core simulator backend-bikesurbanfleets-usersgenerator --> User generator that generates configuration files with single users. 	This module contains some classes of the core too.
This project separation in modules is important, to create an user generator and mantain the code used in the core.
Also the idea of separated projects could help to improve quality and tests.
This branch will be merged with instructions and commands to build the backend and work with them.